# odahuFlow/operator/docs/swagger.yaml
definitions:
...
  Connection:
    properties:
      id:
        description: Connection id
        type: string
      spec:
        $ref: '#/definitions/ConnectionSpec'
        description: Connection specification
        type: object
      status:
        $ref: '#/definitions/ConnectionStatus'
        description: Connection status
        type: object
    type: object
...
  ModelRoute:
    properties:
      id:
        description: Model route id
        type: string
      spec:
        $ref: '#/definitions/ModelRouteSpec'
        description: Model route specification
        type: object
      status:
        $ref: '#/definitions/ModelRouteStatus'
        description: Model route status
        type: object
    type: object
...
  HTTPResult:
    properties:
      message:
        description: Success of error message
        type: string
    type: object
...
  ConnectionSpec:
    properties:
      description:
        description: Custom description
        type: string
      keyID:
        description: Key ID
        type: string
      keySecret:
        description: SSH or service account secret
        type: string
      password:
        description: Password
        type: string
      publicKey:
        description: SSH public key
        type: string
      reference:
        description: VCS reference
        type: string
      region:
        description: AWS region or GCP project
        type: string
      role:
        description: Service account role
        type: string
      type:
        description: |-
          Required value. Available values:
            * s3
            * gcs
            * azureblob
            * git
            * docker
        type: string
      uri:
        description: URI. It is required value
        type: string
      username:
        description: Username
        type: string
      webUILink:
        description: Custom web UI link
        type: string
    type: object
  ConnectionStatus:
    properties:
      secretName:
        description: Kubernetes secret name
        type: string
      serviceAccount:
        description: Kubernetes service account
        type: string
    type: object
...
  ModelDeploymentTarget:
    properties:
      mdName:
        description: Model Deployment name
        type: string
      weight:
        description: The proportion of traffic to be forwarded to the Model Deployment.
        type: integer
    type: object
...
  ModelRouteSpec:
    properties:
      mirror:
        description: |-
          Mirror HTTP traffic to a another Model deployment in addition to forwarding
          the requests to the model deployments.
        type: string
      modelDeployments:
        description: A http rule can forward traffic to Model Deployments.
        items:
          $ref: '#/definitions/ModelDeploymentTarget'
        type: array
      urlPrefix:
        description: |-
          URL prefix for model  For example: /custom/test
          Prefix must start with slash
          "/feedback" and "/model" are reserved for internal usage
        type: string
    type: object
  ModelRouteStatus:
    properties:
      edgeUrl:
        description: Full url with prefix to a model deployment service
        type: string
      state:
        description: State of Model Route
        type: string
    type: object
...
info:
  contact: {}
  description: This is an API Gateway server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Gateway
  version: "1.0"
paths:
...
  /api/v1/connection:
    get:
      consumes:
      - application/json
      description: Get list of Connections
      parameters:
      - description: Toolchain
        in: path
        name: type
        type: string
      - description: Number of entities in a response
        in: path
        name: size
        type: integer
      - description: Number of a page
        in: path
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Connection'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get list of Connections
      tags:
      - Connection
    post:
      consumes:
      - application/json
      description: Create a Connection. Results is created Connection.
      parameters:
      - description: Create a Connection
        in: body
        name: connection
        required: true
        schema:
          $ref: '#/definitions/Connection'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Create a Connection
      tags:
      - Connection
    put:
      consumes:
      - application/json
      description: Update a Connection. Results is updated Connection.
      parameters:
      - description: Update a Connection
        in: body
        name: connection
        required: true
        schema:
          $ref: '#/definitions/Connection'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HTTPResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Update a Connection
      tags:
      - Connection
  /api/v1/connection/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Connection by id
      parameters:
      - description: Connection id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HTTPResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Delete a Connection
      tags:
      - Connection
    get:
      consumes:
      - application/json
      description: Get a Connection by id
      parameters:
      - description: Connection id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get a Connection
      tags:
      - Connection
  /api/v1/connection/{id}/decrypted:
    get:
      consumes:
      - application/json
      description: Get a decrypted Connection by id
      parameters:
      - description: Connection id
        in: path
        name: id
        required: true
        type: string
      - description: Decrypt token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get a decrypted Connection
      tags:
      - Connection
...
  /api/v1/model/route:
    get:
      consumes:
      - application/json
      description: Get list of Model routes
      parameters:
      - description: Number of entities in a response
        in: path
        name: size
        type: integer
      - description: Number of a page
        in: path
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ModelRoute'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get list of Model routes
      tags:
      - Route
    post:
      consumes:
      - application/json
      description: Create a Model route. Results is created Model route.
      parameters:
      - description: Create a Model route
        in: body
        name: mr
        required: true
        schema:
          $ref: '#/definitions/ModelRoute'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ModelRoute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Create a Model route
      tags:
      - Route
    put:
      consumes:
      - application/json
      description: Update a Model route. Results is updated Model route.
      parameters:
      - description: Update a Model route
        in: body
        name: mr
        required: true
        schema:
          $ref: '#/definitions/ModelRoute'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ModelRoute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Update a Model route
      tags:
      - Route
  /api/v1/model/route/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Model route by id
      parameters:
      - description: Model route id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HTTPResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Delete a Model route
      tags:
      - Route
    get:
      consumes:
      - application/json
      description: Get a Model route by id
      parameters:
      - description: Model route id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ModelRoute'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get a Model route
      tags:
      - Route
...
swagger: "2.0"